Changed primary keys of all models to UUID

+++ lib/models/room_rate.ex
@@ -4,19 +4,19 @@ defmodule Auberge.RoomRate do
   import Ecto.Changeset
 
   schema "room_rates" do
-    field :description, :string       # Rack
-    field :code, :string              # RACKMF
-    field :type, :string              # Unused
-    field :starts_at, :date       # 2016-02-01
-    field :ends_at, :date         # 9999-02-01
-    field :days_of_week, :map         # {"Monday": true}
-    field :min_stay, :integer         # 1
-    field :max_stay, :integer         # 7
-    field :min_occupancy, :integer    # 1
-    field :max_occupancy, :integer    # 2
-    field :extra_adult_price, :float  # 199.00
-    field :extra_child_price, :float  # 59.00
-    field :price, :float              # 261.00
+    field :description, :string         # Rack
+    field :code, :string                # RACKMF
+    field :type, :string                # Unused
+    field :starts_at, :date             # 2016-02-01
+    field :ends_at, :date               # 9999-02-01
+    field :days_of_week, :map           # {"Monday": true}
+    field :min_stay, :integer           # 1
+    field :max_stay, :integer           # 7
+    field :min_occupancy, :integer      # 1
+    field :max_occupancy, :integer      # 2
+    field :extra_adult_price, :decimal  # 199.00
+    field :extra_child_price, :decimal  # 59.00
+    field :price, :decimal              # 261.00
 
     timestamps()
     field :deleted_at, :utc_datetime
